upsampleGltfForRasterOverlays -> RasterOverlayUtilities
