+++ Cesium3DTilesSelection/src/LayerJsonTerrainLoader.cpp
@@ -1,6 +1,5 @@
 #include "LayerJsonTerrainLoader.h"
 
-#include <Cesium3DTilesContent/upsampleGltfForRasterOverlays.h>
 #include <CesiumAsync/IAssetResponse.h>
 #include <CesiumGeospatial/calcQuadtreeMaxGeometricError.h>
 #include <CesiumGltfContent/GltfUtilities.h>
@@ -13,7 +12,6 @@
 #include <rapidjson/document.h>
 
 using namespace CesiumAsync;
-using namespace Cesium3DTilesContent;
 using namespace Cesium3DTilesSelection;
 using namespace CesiumGeometry;
 using namespace CesiumGeospatial;
@@ -220,9 +218,7 @@ void generateRasterOverlayUVs(
                                 pParentRegion->getRectangle())
                           : std::nullopt,
             {projection},
-            false,
-            "_CESIUMOVERLAY_",
-            0);
+            false);
   }
 }
 
@@ -1138,11 +1134,11 @@ CesiumAsync::Future<TileLoadResult> LayerJsonTerrainLoader::upsampleParentTile(
        boundingVolume = tile.getBoundingVolume(),
        textureCoordinateIndex = index,
        tileID = *pUpsampledTileID]() mutable {
-        auto model = upsampleGltfForRasterOverlays(
+        auto model = RasterOverlayUtilities::upsampleGltfForRasterOverlays(
             parentModel,
             tileID,
             false,
-            "_CESIUMOVERLAY_",
+            RasterOverlayUtilities::DEFAULT_TEXTURE_COORDINATE_BASE_NAME,
             textureCoordinateIndex);
         if (!model) {
           return TileLoadResult::createFailedResult(nullptr);
